Don't do auto-width for cards, and adjust height at breakpoints
Auto-width fucks up the card heights, so get rid of that shit.
Not sure how it actually causes the issue, best guess is it's doing
something with image widths, getting biggest, and scaling the cards
vertically so the widest image is full-width prior to cropping to 25%
width for the card itself? Whatever the case, it was pretty fucking bad.

For the breakpoint stuff, just trying to make the images less likely to
fight for space with the text content at smaller screen sizes.